agent/requirements.txt
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0
pillow>=10.0.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
openai>=1.0.0
pytesseract>=0.3.10
